+++ tests/testthat/test-default-scoring-rules.R
@@ -44,49 +44,41 @@ test_that("`select_metrics` works as expected", {
 
 
 # ==============================================================================
-# customise_metric()
+# Customising metrics using purrr::partial()
 # ==============================================================================
-
-test_that("customise_metric handles errors correctly", {
-  # Test with a non-function metric
-  expect_error(
-    customise_metric("not_a_function", na.rm = TRUE),
-    "Must be a function, not 'character'"
-  )
-})
 
 test_that("customize_metric is exported", {
   expect_equal(customise_metric, customize_metric)
 })
 
 
-test_that("customise_metric works correctly", {
+test_that("customising metrics via purr::partial works correctly", {
   # Create a customised metric function
-  custom_metric <- customise_metric(mean, na.rm = TRUE)
+  custom_metric <- purrr::partial(mean, na.rm = TRUE)
 
   # Use the customised metric function
   values <- c(1, 2, NA, 4, 5)
   expect_equal(custom_metric(values), 3)
 
   # Test with a different metric function
-  custom_metric <- customise_metric(sum, na.rm = TRUE)
+  custom_metric <- purrr::partial(sum, na.rm = TRUE)
   expect_equal(custom_metric(values), 12)
 
   # Test with no additional arguments
-  custom_metric <- customise_metric(mean)
+  custom_metric <- purrr::partial(mean)
   expect_true(is.na(custom_metric(values)))
 
   # make sure that customise_metric fails immediately (instead of at runtime)
   # when object doesn't exist
   expect_error(
-    custom_metric <- customise_metric(print, x = doesnotexist),
+    custom_metric <- purrr::partial(print, x = doesnotexist),
     "object 'doesnotexist' not found"
   )
 
   # make sure that customise_metric still works even if original object is
   # deleted, meaning that the object is stored as part of the function
   argument <- c("hi", "hello", "I'm here")
-  custom_metric <- customise_metric(print, x = argument)
+  custom_metric <- purrr::partial(print, x = argument)
   expect_output(custom_metric(), "I'm here")
 
   argument <- NULL
@@ -102,8 +94,8 @@ test_that("customise_metric works correctly", {
   )
 })
 
-test_that("customise_metric() has the expected output class", {
-  custom_metric <- customise_metric(mean, na.rm = TRUE)
+test_that("purrr::partial() has the expected output class", {
+  custom_metric <- purrr::partial(mean, na.rm = TRUE)
   checkmate::expect_class(custom_metric, "function")
 })